removed some comments and some small fixes

+++ android/app/src/main/cpp/minis_jni.c
@@ -3,11 +3,11 @@
 #include <stdlib.h>
 #include <pthread.h>
 #include <unistd.h>
-#include <time.h> // Required for clock_gettime
+#include <time.h>
 
-#include "cache/minis.h" // Your Core Public API
+#include "cache/minis.h"
 #include "cache/t_string.h"
-#include "cache/hash.h" // Ensure you include hash.h for HashEntry definition
+#include "cache/hash.h"
 #include "common/macros.h"
 #include "cache/t_hash.h"
 
@@ -37,10 +37,9 @@ static ALWAYS_INLINE uint64_t
 get_us (void) {
     struct timespec tvs = { 0, 0 };
     clock_gettime (CLOCK_MONOTONIC, &tvs);
-    return (uint64_t) ((tvs.tv_sec * 1000000) + (tvs.tv_nsec / 1000));
+    return ((tvs.tv_sec * 1000000) + (tvs.tv_nsec / 1000));
 }
 
-// --- 2. The Safe Background Thread ---
 static void* maintenance_thread(void* arg) {
     MinisContext *ctx = (MinisContext*)arg; // We receive the Context, not just Minis
     // Check the flag on every loop iteration
@@ -62,27 +61,20 @@ static void* maintenance_thread(void* arg) {
     return NULL;
 }
 
-// --- JNI Implementation ---
-
 JNIEXPORT jlong JNICALL
 Java_com_minis_Minis_nativeInit(JNIEnv *env, jobject thiz) {
     (void)env;
     (void)thiz;
-    // 1. Allocate the Context Wrapper
     MinisContext *ctx = malloc(sizeof(MinisContext));
 
-    // 2. Initialize Core
     ctx->minis = minis_init();
     ctx->is_running = 1;
 
-    // 3. Spawn Background Thread (Pass 'ctx' as argument)
-    // We DO NOT detach it, because we need to join (wait) for it later.
     if (pthread_create(&ctx->bg_thread, NULL, maintenance_thread, ctx) != 0) {
         LOGE("Failed to create background thread");
         // Handle error cleanup if needed
     }
 
-    // 4. Return pointer to the CONTEXT (not the engine directly)
     return (jlong) ctx;
 }
 
@@ -94,23 +86,15 @@ Java_com_minis_Minis_nativeFree(JNIEnv *env, jobject thiz, jlong ptr) {
     if (ctx) {
         LOGI("Stopping Minis...");
 
-        // 1. Tell thread to stop
         ctx->is_running = 0;
-
-        // 2. WAIT for thread to actually finish (Critical Fix)
-        // This ensures minis_evict() is never called after we destroy the lock below.
         pthread_join(ctx->bg_thread, NULL);
-
-        // 3. Now it is safe to destroy the core engine
         minis_free(ctx->minis);
 
-        // 4. Free the wrapper
         free(ctx);
         LOGI("Minis Stopped.");
     }
 }
 
-// --- Helper Macro to Unwrap Context ---
 #define GET_MINIS(ptr) (((MinisContext*)(ptr))->minis)
 
 JNIEXPORT void JNICALL
@@ -127,7 +111,6 @@ Java_com_minis_Minis_nativeSet(JNIEnv *env, jobject thiz, jlong ptr, jstring jKe
     (*env)->ReleaseStringUTFChars(env, jVal, val);
 }
 
-// New Context: holds a temporary C-buffer
 typedef struct {
     char *value_copy; // We will malloc here inside lock
 } JniGetCtx;
@@ -137,8 +120,6 @@ jni_get_visitor_fast(const Entry *ent, void *ctx)
 {
     JniGetCtx *jctx = (JniGetCtx *)ctx;
     if (ent && ent->val) {
-        // Fast copy inside lock. 
-        // Much faster than JNI NewStringUTF.
         jctx->value_copy = strdup(ent->val); 
     }
     return true;
@@ -202,7 +183,7 @@ Java_com_minis_Minis_nativePexpire(JNIEnv *env, jobject thiz, jlong ptr, jstring
 JNIEXPORT jlong JNICALL
 Java_com_minis_Minis_nativePttl(JNIEnv *env, jobject thiz, jlong ptr, jstring jKey) {
     (void) thiz;
-    Minis *minis = GET_MINIS(ptr); // Unwrap
+    Minis *minis = GET_MINIS(ptr);
     const char *key = (*env)->GetStringUTFChars(env, jKey, 0);
 
     int64_t ttl = 0;
@@ -215,7 +196,7 @@ Java_com_minis_Minis_nativePttl(JNIEnv *env, jobject thiz, jlong ptr, jstring jK
 JNIEXPORT jboolean JNICALL
 Java_com_minis_Minis_nativeSave(JNIEnv *env, jobject thiz, jlong ptr, jstring jPath) {
     (void) thiz;
-    Minis *minis = GET_MINIS(ptr); // Unwrap
+    Minis *minis = GET_MINIS(ptr);
     const char *path = (*env)->GetStringUTFChars(env, jPath, 0);
 
     MinisError err = minis_save(minis, path, get_us());
@@ -227,7 +208,7 @@ Java_com_minis_Minis_nativeSave(JNIEnv *env, jobject thiz, jlong ptr, jstring jP
 JNIEXPORT jboolean JNICALL
 Java_com_minis_Minis_nativeLoad(JNIEnv *env, jobject thiz, jlong ptr, jstring jPath) {
     (void) thiz;
-    Minis *minis = GET_MINIS(ptr); // Unwrap
+    Minis *minis = GET_MINIS(ptr);
     const char *path = (*env)->GetStringUTFChars(env, jPath, 0);
 
     MinisError err = minis_load(minis, path, get_us());
@@ -264,36 +245,29 @@ jni_mget_visitor(const Entry *ent, void *ctx) {
     return true; 
 }
 
-// --- JNI MGET Implementation ---
-
 JNIEXPORT jobjectArray JNICALL
 Java_com_minis_Minis_nativeMGet(JNIEnv *env, jobject thiz, jlong ptr, jobjectArray jKeys) {
     (void) thiz;
-    Minis *minis = GET_MINIS(ptr); // (Using the context wrapper from before)
+    Minis *minis = GET_MINIS(ptr);
     jsize count = (*env)->GetArrayLength(env, jKeys);
 
-    // 1. Prepare C Keys (same as before)
     const char **cKeys = malloc(sizeof(char*) * count);
     for (int i = 0; i < count; i++) {
         jstring js = (jstring) (*env)->GetObjectArrayElement(env, jKeys, i);
         cKeys[i] = (*env)->GetStringUTFChars(env, js, 0);
     }
 
-    // 2. Prepare Result Array
     jclass stringClass = (*env)->FindClass(env, "java/lang/String");
     jobjectArray result = (*env)->NewObjectArray(env, count, stringClass, NULL);
 
-    // 3. Prepare Visitor Context
     JniMGetContext visitorCtx = {
             .env = env,
             .jArray = result,
             .index = 0
     };
 
-    // 4. Call Core (Uses your new Visitor API!)
     minis_mget(minis, cKeys, count, jni_mget_visitor, &visitorCtx, get_us());
 
-    // 5. Cleanup Keys
     for (int i = 0; i < count; i++) {
         jstring js = (jstring) (*env)->GetObjectArrayElement(env, jKeys, i);
         (*env)->ReleaseStringUTFChars(env, js, cKeys[i]);
@@ -310,7 +284,6 @@ Java_com_minis_Minis_nativeMDel(JNIEnv *env, jobject thiz, jlong ptr, jobjectArr
     Minis *minis = GET_MINIS(ptr);
     jsize count = (*env)->GetArrayLength(env, jKeys);
 
-    // 1. Allocate for Keys AND Refs
     const char **cKeys = malloc(sizeof(char*) * count);
     jstring *jStrings = malloc(sizeof(jstring) * count);
 
@@ -336,19 +309,16 @@ Java_com_minis_Minis_nativeMDel(JNIEnv *env, jobject thiz, jlong ptr, jobjectArr
         successCount++;
     }
 
-    // 2. Call Core
     uint64_t deleted = 0;
     if (!failed) {
         minis_mdel(minis, cKeys, count, &deleted, get_us());
     }
 
-    // 3. Cleanup
     for (int i = 0; i < successCount; i++) {
         (*env)->ReleaseStringUTFChars(env, jStrings[i], cKeys[i]);
         (*env)->DeleteLocalRef(env, jStrings[i]);
     }
 
-    // Clean up the partial failure item if it exists
     if (failed && jStrings[successCount] != NULL) {
         (*env)->DeleteLocalRef(env, jStrings[successCount]);
     }
@@ -367,12 +337,9 @@ Java_com_minis_Minis_nativeMSet(JNIEnv *env, jobject thiz, jlong ptr, jobjectArr
 
     if (count % 2 != 0) return;
 
-    // 1. Allocate arrays for C strings AND Java string references
-    // We MUST save jStrings to release them correctly later.
     const char **cKeyVals = malloc(sizeof(char*) * count);
     jstring *jStrings = malloc(sizeof(jstring) * count);
 
-    // Safety: OOM check
     if (!cKeyVals || !jStrings) {
         if(cKeyVals) free(cKeyVals);
         if(jStrings) free(jStrings);
@@ -383,14 +350,11 @@ Java_com_minis_Minis_nativeMSet(JNIEnv *env, jobject thiz, jlong ptr, jobjectArr
     int successCount = 0;
 
     for (int i = 0; i < count; i++) {
-        // Get and SAVE the reference
         jstring js = (jstring) (*env)->GetObjectArrayElement(env, jKeyVals, i);
         jStrings[i] = js;
 
-        // Convert
         cKeyVals[i] = (*env)->GetStringUTFChars(env, js, 0);
 
-        // Safety: Check for NULL (JNI OOM)
         if (cKeyVals[i] == NULL) {
             failed = true;
             break;
@@ -398,19 +362,15 @@ Java_com_minis_Minis_nativeMSet(JNIEnv *env, jobject thiz, jlong ptr, jobjectArr
         successCount++;
     }
 
-    // 2. Call Core (Only if conversion succeeded)
     if (!failed) {
         minis_mset(minis, cKeyVals, (size_t)count, get_us());
     }
 
-    // 3. Cleanup using the SAVED references
-    // We only clean up what we successfully allocated
     for (int i = 0; i < successCount; i++) {
         (*env)->ReleaseStringUTFChars(env, jStrings[i], cKeyVals[i]);
         (*env)->DeleteLocalRef(env, jStrings[i]); // Now we delete the saved ref
     }
 
-    // If we failed in the middle, we still need to release the jstring causing the failure
     if (failed && jStrings[successCount] != NULL) {
         (*env)->DeleteLocalRef(env, jStrings[successCount]);
     }
@@ -447,13 +407,11 @@ Java_com_minis_Minis_nativeHSet(JNIEnv *env, jobject thiz, jlong ptr, jstring jK
     return (jint)added;
 }
 
-// Struct to hold context for HGET visitor
 typedef struct {
     JNIEnv *env;
     jstring result;
 } JniHGetCtx;
 
-// Specific Visitor for HGET (Single Field)
 static bool jni_hget_visitor(const HashEntry *entry, void *arg) {
     JniHGetCtx *ctx = (JniHGetCtx *)arg;
     if (entry && entry->value) {
@@ -578,10 +536,8 @@ static bool jni_hgetall_visitor(const HashEntry *entry, void *arg) {
     jstring jField = (*ctx->env)->NewStringUTF(ctx->env, entry->field);
     jstring jValue = (*ctx->env)->NewStringUTF(ctx->env, entry->value);
 
-    // map.put(field, value)
     (*ctx->env)->CallObjectMethod(ctx->env, ctx->hashMap, ctx->putMethod, jField, jValue);
 
-    // Clean up local references to prevent JNI table overflow in large loops
     (*ctx->env)->DeleteLocalRef(ctx->env, jField);
     (*ctx->env)->DeleteLocalRef(ctx->env, jValue);
